refactor: update include paths for detected object filters
