+++ perception/detected_object_validation/src/lanelet_filter/lanelet_filter.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "detected_object_validation/detected_object_filter/object_lanelet_filter.hpp"
+#include "lanelet_filter.hpp"
 
 #include <autoware/universe_utils/geometry/geometry.hpp>
 #include <autoware_lanelet2_extension/utility/message_conversion.hpp>
