+++ perception/detected_object_validation/src/position_filter/position_filter.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "detected_object_validation/detected_object_filter/object_position_filter.hpp"
+#include "position_filter.hpp"
 
 namespace object_position_filter
 {
