+++ perception/detected_object_validation/src/utils/utils.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "detected_object_validation/utils/utils.hpp"
+#include "utils.hpp"
 
 #include <autoware_perception_msgs/msg/object_classification.hpp>
