+++ perception/detected_object_validation/src/utils/utils.hpp
