+++ perception/detected_object_validation/test/object_position_filter/test_object_position_filter.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "detected_object_validation/detected_object_filter/object_position_filter.hpp"
+#include "../../src/position_filter/position_filter.hpp"
 
 #include <ament_index_cpp/get_package_share_directory.hpp>
 #include <autoware_test_utils/autoware_test_utils.hpp>
